requirements.txt
Django==2.0.1
djangorestframework==3.7.7
Pillow==5.0.0
pkg-resources==0.0.0
pytz==2017.3
